Modernize supervision and app
